requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.4
makefile==1.1.0
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
Faker==30.8.2
environs==11.0.0